wm/requirement.txt
python-xlib
pillow
